--- src/old_src/includeOggVorbis/Makefile.am
@@ -1,3 +1,0 @@
-## Process this file with automake to produce Makefile.in
-
-SUBDIRS = vorbis
