--- src/old_src/includeOggVorbis/ogg/Makefile.am
@@ -1,6 +1,0 @@
-## Process this file with automake to produce Makefile.in
-
-oggincludedir = $(includedir)/ogg
-
-ogginclude_HEADERS = ogg.h os_types.h
-nodist_ogginclude_HEADERS = config_types.h
