--- src/old_src/includeOggVorbis/vorbis/Makefile.am
@@ -1,7 +1,0 @@
-## Process this file with automake to produce Makefile.in
-
-vorbisincludedir = $(includedir)/vorbis
-
-vorbisinclude_HEADERS = codec.h vorbisfile.h vorbisenc.h
-
-
