+++ src/src/diary.txt
@@ -65,3 +65,7 @@
            : これでTonyu1(DirectX6以降)が動く環境なら動かせるようになった
 2019/10/27 : PCを１分以上スリープした後復帰すると、Midiが勝手に再生されるバグを修正
            : 何日もスリープした後GuruguruSMFが固まる問題に対処する処理が、中途半端に入っていた
+2020/12/12 : 新しく作り直し
+           : ライブラリ更新。libogg 1.3.3 -> 1.3.4 , libvorbis 1.3.6 -> 1.3.7
+
+
